feat: 🎸 Add fullExtract parameter to exportingCsvDeltaFile
If event.fullExtract is true then set the since date to well in the
past, so we get all the current records, and also ignore any deletes
function. For a full export, we want the current state of the tables, so
any deletes that might appear in an incremental export SHOULD NOT appear
in a full extract.